TOOLS:  Added packtool, which packages files in EMBED with app
The packtool allows an application to package arbitrary files within it, circumventing the limitations of some
platform (notably iOS and Android) packaging schemes. It is used to include configuration files.

The tool looks for an EMBED file with space separated filenames relative to the current directory. If the EMBED file
exists the listed files will be compressed and put into to scheme file embed.scm. For this tool to work  (define
"embed.scm")  needs to be present at the top of an application's main.scm file, which then will automatically unpack
the files into the system directory. Unpacking happens only at the first launch of the application, so files don't get
overwritten.