Compilation and permissions for ANDROIDAPI>22. (#257)
Since Marshmallow permissions are not automatically granted as specified in the Manifest. As a consequence LN apps crash when accessing the SDCard, e.g. when unpacking embedded files.

This patch adds a fragile hack to enable an ifdef-alike trick to exclude code portions when targeting older APIs and uses it to request the permission at startup.  It also creates the system-directory if it does not already exists.