Err on usability wrt. clipboard content.
On linux and wind32 the code does so far not know wheter or not the
clipboard has any content.

That's NOT changed by this patch.

However it's IMHO better to provide a false positive claiming
content (thus suggesting to the user to paste is even though that
would fail) than falsely claim nothing available thus preventing avail
content to be pasted.