LN_CORE: allow re-initializing logging at runtime (#272)
New procedure (log-reset! . file) may be called to re-initialize logging. If the logdir has been removed, logging remains disabled. If it was created in the meantime, logging is started.

In any case, logging is switched to either:
a) a new file using the naming convention as before; starting at the (current-time)
b) or into the file given as optional parameter to `log-reset!`