LNjScheme Demo App (Android only) (#273)
This is a demo how a simple Scheme interpreter in Java (derived from Peter Norvig's jScheme) can be used to run arbitrary Jave/Scheme code in either the main UI thread of Android or any other thread. Due to Android's requirements with respect to changing views the former is likely the more interesting option.

The main idea behind is to generalize the special cases currently used for hybridapps (and remove the special handling from the build process), while making it easier to use native widgets from lambdanative. For further details, please see the README.md file in the app's folder.