LN_GLCORE: Resolve texture rendering problem on MacOS/iOS from _glCoreTextureCreate (closes #297)
The issue is that _glCoreTextureCreate sometimes does not properly set up the texture, but marks it as initialized anyway. This fix adds a further check on the proper operation of the GL Context, creates an error, and lets the texture be re-initialized on next call. This stems from being called in a different thread from the main GL Context. This also fixes what I believe to be a typo in _glCoreTextureBind, and adds helpful functions glGetError and glIsEnabled to glcore-ffi.scm. Thanks @peterlew