GAMBIT: Supplement to 7cbb2c3: precompiled C file.
7cbb2c3 updates the gambit source.  To install normally the
precompiled file _io.c is required to be up to date wrt _io.scm.

This patch contains this pre-compiled file.

It also adds a single line to libraries/libgambit/make.sh, which
mimicks the skipped parts of a normal gambit build.  This might be
useful in case the build finds the _io.c file outdated.

In order to not break the workflow when updating lambdanative and keep
this patch readable, there will be another one submitting the
precompiled _io.c file.  Both patches should be either both or none of
them applied.