HYBRIDAPP: Android APIs>24 no longer honor user CAs unless forced
This resolves a problem, whereby the webview will not open custom
URLs for valid https certificate chains if the root is not in the
system certificate pool. This commit trusts all user-added CAs
for all secure connections, but might better be restricted to
for some domains. See https://android-developers.googleblog.com/
2016/07/changes-to-trusted-certificate.html for details.

The modules/hybridap/ANDROID_application_attributes @xml is
fragile so if editing it may require a second line for subtool to
not stop prematurely.