SCM: add target specific directory definition to gambit invocation
Gambit allows to refer to various installation dependent directories
using the ~~NAME symtax - e.g., as in `(include "~~lib/my/file.scm")`.
However gambit is not aware of cross-compiling.  So `~~lib` will refer
to the hosts version.

This change adds another option to the gambit invocation, which allows
to distinguish between the host and the target specific files of
gambit.  At least for the library.

The use case this is required for is not (yet) publically available.
It involves a library wich includes gambit source code using
cond-expand to include use cases like file locking for instance -
which is a horror to provide in a platform agnostic way.