GAMBIT: improve IPv6 handling on the client side (#320)
This patch contains this large pre-compiled file patch for _io.c. This is done to avoid having to go through the bootstrapping process of gambit needed otherwise to apply the change to lib/os_io.c. It also adds a single line to libraries/libgambit/make.sh, which mimics the skipped parts of a normal gambit build in case the build finds the _io.c file outdated.