SCM.SH: add target directory definition to gambit invocation (#321)
Gambit allows to refer to various installation dependent directories using the ~~NAME syntax - e.g., as in (include "~~lib/my/file.scm"). However, gambit is not aware of cross-compiling. So ~~lib will refer to the hosts version. This adds an option to the gambit invocation, which allows distinguishing between the host and the target specific files of
gambit.