Article: Summary Of First Assessment
Huston: we's got a problem.

The bad news: Complexity issues are worth than load.

The good news: It is small in practice.  Because our keybords have
only so many columns and rows and while a complexity of O(n²) is
horror in principle because it does not scale, our keybords do not
grow so fast as our processors become faster.

The worst news so far: It's late today.  The mind just does not want
to try to quickly fix that.  We have:

1.) code raising an exception

2.) Installing exception handler is relative expensive.  Given the
task there should be no need to fail in list-ref.

3.) list-ref is O(n) AND directly AND indirectly called within a loop.

Question: can we understand the algorith to the extent that we can do
without list-ref at all?