GLGUI: reduce idle load (#330)
This patch is a fundamental change of how frame rates are determined on desktop platforms. For mobile platforms: iOS and android, as well as any applications using the serial and scheduler modules there should be no difference in their timing (max of 100 Hz). This approach results in reduced CPU load, but it also makes the frame rate vary between 2 and 20 Hz. The delay increases increasing numbers of consecutive 'EVENT_REDRAW's resulting in the maximum frame rate to decrease from initially 20 Hz to 2 Hz in 1 Hz steps.

Note: When other threads notice that the GUI thread should stop waiting and thus reducing responsiveness : call (glgui-wakeup!).