GLGUI: Revise chat widget caching to reduce draw complexity (#350)
Previously, the chat widget cached a pre-processed version of the list of entries to display, focusing on the 'length' of the list. This change slightly modifies the cache to improve scaling for longer lists. New caching tests ONLY, whether or not the old list is eq? the new one. It does NOT try to detect destructive updates to the components of the list.