GLGUI: reduce complexity of font operations. (#359)
Note: In order to keep the patch reasonably sized, this patch is only a first of a serious of changes. So far it only proofs the concept and cracks down on the most prominent cases profiling the app revealed. Over time the use of 'assoc' on fonts should be phased out and replaced with properly abstracted operations to enable the underlying implementation to be changed.
Note2: This change overwrites assoc globally, but should do so in a benign manner!