CORE: optimize time formatting, fix typo and wrong comment
 # Background

As I've seen this quasi-mistake often, I fell compelled to explain.

The code this is really, really code calling for two steps of rewite
for otimization:

    (define (string->seconds str fmt . tz0)
      (apply string->seconds~ (append (list str (ln:tildify fmt)) tz0)))

 ## Step 1

apply usually (and with gambit) takes possibly more arguments.  So

    (apply f e1 e2 ... lst)
is the same as

    (apply f (append (list e1 e2 ...) lst))

Therefore we better write:

    (define (string->seconds str fmt . tz0)
      (apply string->seconds~ str (ln:tildify fmt) tz0))

 ## Step 2

`apply` is usually less effective as a direct call.  For frequently
called code this counts.

Therefore we write better dispatch on tz0:

(define (string->seconds str fmt . tz0)
  (if (pair? tz0)
      (string->seconds~ str (ln:tildify fmt) (car tz0))
      (string->seconds~ str (ln:tildify fmt))))

This has still the same effect as now in the first arm of the `if`
expression the passing of the tz0 reaches `string->seconds~` and in
the second arm no rest argument is passed.  Exactly the same effect as
the version using `apply`.

The comment "can't do optimization here" is therefore not really correct.