ANDROID: additional error checking on JNI
Android 9/10 restricts access to JNI and Jave reflection API's.

This adds additional checks.

Commented out an possible optimization, which could be assumed to
works, but does not for me so far.  Please leave this comment in, so
we recall before we wonder why and try over and over again.  Once in a
while we should ponder if it still does not work or why.