ANDROID: additional error checking on JNI an more
1.) Android 9/10 restricts access to JNI and Jave reflection API's.

This adds additional checks.

Commented out an possible optimization, which could be assumed to
works, but does not for me so far.  Please leave this comment in, so
we recall before we wonder why and try over and over again.  Once in a
while we should ponder if it still does not work or why.

2.) Clear Java exceptions.

This enables to continue to run when an exception occured during a JNI
call.  (Future versions shall forward this as an exception in Gambit.)

3.) Avoid some JNI calls and provide more information about the
Java/Android environment to Scheme.

This is required for (upcoming) tricks to still call embedded dynamic
libraries as subprocesses.

It also enables to figure out a sane path to store app-private data
instead of the deprecated hard coding of the publicly accessible path
`/sdcard`.

4.) Add support to switch the apps content view to Java and back.

An upcoming module `webview` will need this.