LANGUAGES: Fix silent segfault due to module dependency in #330
Using "(not safe)" in languages/scm.sh causes a segfault whenever apps include scheduler before ln_glgui in their MODULE list.

This was extremely hard to debug as there is no crash dump, lldb only shows a jump error and it doesn't happen in debug mode. In the end the actual problem is (#!unbound2 frame-period-custom: '#<procedure #2>) caused by (glgui-timings-at-10msec!) being called in a function-exists? in modules/scheduler/scheduler.scm, while glgui-timings-set! is an internally declared function pointing to timings-set!, which is only initialized after the define of glgui-event runs.

The solution is to either ensure scheduler is never listed before ln_glgui in MODULES, or fixing this by making it a scm_coremodules, which this change does. It is possible that this order change causes other problems for custom modules that need to be run before ln_glgui, but are not a scm_coremodules but the ones we ship are fine.