MICROGL X11: Switch from X11 Expose events to timer-based renderer
A patch contributed by @zwieblum that makes the framerate stable and predictable at approximately 50Hz - just like you would need for games. Set (glgui-timings-at-sec! 0.0) or any value smaller 16 msec to disable #330. If you don't, the behaviour is widely unchanged, but the event queue is not saturated with Expose events. Closes #444.